pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cofounder"
version = "0.1.0"
description = "AI co-founder for solo builders — a persistent, opinionated thinking partner"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.30.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "websockets>=12.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
cofounder = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
